requirements.txt
# =============================================================================
# Nickelfront - Unified Requirements (Python 3.13+)
# =============================================================================

# -----------------------------------------------------------------------------
# API
# -----------------------------------------------------------------------------
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.30.0

# -----------------------------------------------------------------------------
# Database
# -----------------------------------------------------------------------------
sqlalchemy>=2.0.36
asyncpg>=0.30.0
alembic>=1.14.0
aiosqlite>=0.20.0
psycopg2-binary>=2.9.10

# -----------------------------------------------------------------------------
# Queue
# -----------------------------------------------------------------------------
celery>=5.4.0
redis>=5.2.0
flower>=2.0.1

# -----------------------------------------------------------------------------
# Core
# -----------------------------------------------------------------------------
pydantic>=2.10.0
pydantic-settings>=2.6.0

# -----------------------------------------------------------------------------
# Security
# -----------------------------------------------------------------------------
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# -----------------------------------------------------------------------------
# HTTP Client (for parsers)
# -----------------------------------------------------------------------------
httpx>=0.27.0
requests>=2.32.0

# -----------------------------------------------------------------------------
# Logging
# -----------------------------------------------------------------------------
loguru>=0.7.2

# -----------------------------------------------------------------------------
# Testing
# -----------------------------------------------------------------------------
pytest>=8.3.0
pytest-asyncio>=0.24.0

# -----------------------------------------------------------------------------
# RAG (Retrieval-Augmented Generation)
# -----------------------------------------------------------------------------
langchain>=0.3.0,<0.4.0
langchain-community>=0.3.0,<0.4.0
langchain-core>=0.3.0,<0.4.0
langchain-text-splitters>=0.3.0,<0.4.0
pdfplumber>=0.11.0
python-multipart>=0.0.20

# -----------------------------------------------------------------------------
# Vector Search & ML
# -----------------------------------------------------------------------------
chromadb>=1.5.5
sentence-transformers>=3.0.0

# -----------------------------------------------------------------------------
# Analytics & Data Processing
# -----------------------------------------------------------------------------
polars>=1.0.0
pandas>=2.2.0
numpy>=1.26.0

# -----------------------------------------------------------------------------
# Reports & Export
# -----------------------------------------------------------------------------
reportlab>=4.2.0
python-docx>=1.1.0
weasyprint>=62.0

# -----------------------------------------------------------------------------
# Qwen Service Integration
# -----------------------------------------------------------------------------
# Qwen API client (локальный пакет)
# Устанавливается через: pip install -e qwen_service/

# -----------------------------------------------------------------------------
# Auto-detected dependencies (Python 3.13 compatible)
# -----------------------------------------------------------------------------

# Celery Deps
amqp>=5.3.0
billiard>=4.2.0
kombu>=5.4.0
vine>=5.1.0

# HTTP Stack
anyio>=4.6.0
asgiref>=3.8.0
certifi>=2024.0.0
charset-normalizer>=3.4.0
h11>=0.14.0
httpcore>=1.0.0
httptools>=0.6.0
idna>=3.10
oauthlib>=3.2.0
sniffio>=1.3.0
urllib3>=2.2.0
websocket-client>=1.8.0
websockets>=13.0.0

# Crypto/Security
bcrypt>=4.0.1,<5.0.0
cffi>=1.17.0
cryptography>=43.0.0
ecdsa>=0.19.0
pyasn1>=0.6.0
rsa>=4.9.0

# Parsing/HTML/PDF
cssselect2>=0.7.0
fonttools>=4.54.0
lxml>=5.3.0
pydyf>=0.10.0
pyphen>=0.16.0
tinycss2>=1.3.0
tinyhtml5>=2.0.0
webencodings>=0.5.1
zopfli>=0.2.0

# Chroma/Vector
mmh3>=5.0.0

# ML/AI
huggingface_hub>=0.26.0
joblib>=1.4.0
mpmath>=1.3.0
networkx>=3.4.0
onnxruntime>=1.19.0
regex>=2024.0.0
safetensors>=0.4.0
scikit-learn>=1.5.0
scipy>=1.14.0
sympy>=1.13.0
threadpoolctl>=3.5.0
tokenizers>=0.20.0
torch>=2.5.0
tqdm>=4.66.0
transformers>=4.46.0

# Observability
coloredlogs>=15.0.0
humanfriendly>=10.0.0
humanize>=4.9.0
opentelemetry-api>=1.27.0
opentelemetry-exporter-otlp-proto-common>=1.27.0
opentelemetry-exporter-otlp-proto-grpc>=1.27.0
opentelemetry-instrumentation>=0.48b0
opentelemetry-instrumentation-asgi>=0.48b0
opentelemetry-instrumentation-fastapi>=0.48b0
opentelemetry-proto>=1.27.0
opentelemetry-sdk>=1.27.0
opentelemetry-semantic-conventions>=0.48b0
opentelemetry-util-http>=0.48b0
posthog>=4.0.0
prometheus_client>=0.21.0

# CLI/UX
click>=8.1.0
click-didyoumean>=0.3.0
click-plugins>=1.1.0
click-repl>=0.3.0
colorama>=0.4.6
prompt_toolkit>=3.0.0
pyreadline3>=3.5.0
rich>=13.9.0
shellingham>=1.5.0
typer>=0.12.0
wcwidth>=0.2.0

# Utilities
annotated-doc>=0.0.0
annotated-types>=0.7.0
backoff>=2.2.0
build>=1.2.0
distro>=1.9.0
durationpy>=0.9
filelock>=3.16.0
flatbuffers>=24.0.0
fsspec>=2024.0.0
googleapis-common-protos>=1.65.0
importlib_metadata>=8.5.0
importlib_resources>=6.4.0
iniconfig>=2.0.0
Jinja2>=3.1.0
markdown-it-py>=3.0.0
MarkupSafe>=3.0.0
mdurl>=0.1.0
orjson>=3.10.0
overrides>=7.7.0
packaging>=24.0.0
pillow>=10.4.0
pluggy>=1.5.0
polars-runtime-32>=1.0.0
pyproject_hooks>=1.2.0
python-dateutil>=2.9.0
python-dotenv>=1.0.0
pytz>=2024.0.0
PyYAML>=6.0.0
six>=1.16.0
tenacity>=9.0.0
tomli>=2.0.0
typing_extensions>=4.12.0
tzdata>=2024.0.0
watchfiles>=0.24.0
win32_setctime>=1.2.0
wrapt>=1.16.0
zipp>=3.20.0

# Other
async-timeout>=5.0.0
brotli>=1.1.0
dnspython>=2.6.0
email-validator>=2.2.0
greenlet>=3.1.0
grpcio>=1.66.0
kubernetes>=31.0.0
Mako>=1.3.0
protobuf>=5.28.0
pycparser>=2.22
pydantic_core>=2.27.0
Pygments>=2.18.0
PyPika>=0.51.0
starlette>=0.41.0
tornado>=6.4.0
